added Criteria API based mathod to ReportRepository

+++ src/main/java/dev/akuniutka/cbrratesbot/dto/CountDto.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class CountDto {
-    private final int count;
+    private final long count;
 }
